moving languageIndexingConfiguration into new file and break down the code
